+++ src/components/ActionModal/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "ActionModal.js",
+  "version": "0.0.0",
+  "main": "./ActionModal.js"
+}
