--- src/components/SWActivationModal/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "SWActivationModal.js",
-  "version": "0.0.0",
-  "main": "./SWActivationModal.js"
-}
