+++ storybook/storyLoader.js
@@ -4,13 +4,13 @@
 // https://github.com/elderfo/react-native-storybook-loader.git
 
 function loadStories() {
+	require('../src/components/ActionModal/ActionModal.stories');
 	require('../src/components/InsightWithButton/InsightWithButton.stories');
-	require('../src/components/SWActivationModal/SWActivationModal.stories');
 }
 
 const stories = [
-	'../src/components/InsightWithButton/InsightWithButton.stories',
-	'../src/components/SWActivationModal/SWActivationModal.stories'
+	'../src/components/ActionModal/ActionModal.stories',
+	'../src/components/InsightWithButton/InsightWithButton.stories'
 ];
 
 module.exports = {
